requirements.txt
Django==3.0.4
pdfkit==0.6.1
androguard==3.3.5
lxml==4.5.0
rsa==4.0
biplist==1.0.3
requests==2.23.0
bs4==0.0.1
colorlog==4.1.0
macholib==1.14
google-play-scraper==0.0.2.2
whitenoise==5.0.1
waitress==1.4.3
gunicorn==20.0.4
frida==12.8.14
psutil==5.7.0
shelljob==0.5.6
asn1crypto==1.3.0
oscrypto==1.1.0
distro==1.4.0
IP2Location==8.4.1
# Do not update the following
apkid==2.1.0
cryptography<2.5,>=2.1.4 # pyup: ignore
pyOpenSSL==19.0.0 # pyup: ignore
http-tools==1.1.0
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-formtools==2.2
django-import-export==2.3.0
django-redis==4.12.1
django-reversion==3.0.8
PyJWT==1.7.1
PyMySQL==0.10.0
aiohttp==3.6.2
httplib2==0.18.1
